feat(dashboard): implement DashboardHeader, DashboardLayoutClient, and DashboardSidebar components; add navigation items and page skeleton for improved layout and user experience
